requirements.txt
# SCOPE3 - FastAPI Requirements
# Optimized for production deployment on Render.com

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Templates & Static Files
jinja2==3.1.3
aiofiles==23.2.1

# Database
supabase==2.3.0
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client (async)
httpx==0.26.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Financial Data
yfinance==0.2.35
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1

# AI Integration
google-generativeai==0.3.2

# Audio (TTS)
edge-tts==6.1.10
nest-asyncio==1.6.0

# Utilities
requests==2.31.0
openpyxl==3.1.2
